+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 This project does not follow SemVer, since modules are independent of each other; thus, SemVer does not make sense. Changes are grouped per module.
 
 ## Unreleased
+### circleci-token-rotator
+- New module to rotate IAM user's AWS Access Keys on CircleCI. [#278](https://github.com/dbl-works/terraform/pull/278)
+
 ### stack/app
 - Set default rds_engine_version to 16.
 
@@ -18,6 +21,7 @@ This project does not follow SemVer, since modules are independent of each other
 
 ### ecs
 - Add enable_container_insights variables. [#282](https://github.com/dbl-works/terraform/pull/282)
+
 
 ## [v2023.12.03]
 ### ecs